+++ lib/events/default.rb
@@ -1,3 +1,5 @@
+require_relative '../logic'
+
 SlackRubyBotServer::Events.configure do |config|
   config.on :event do |event|
     event.logger.info "#{event[:event][:user]}"
@@ -9,28 +11,5 @@ SlackRubyBotServer::Events.configure do |config|
     end
 
     { ok: true }
-  end
-end
-
-class BotUser
-  def say(msg)
-    if msg_valid?(msg)
-      add_from(msg)
-    else
-      "Sorry, I didn't understand that. I only add numbers in this format. eg: 5+6 or 6+3"
-    end
-  end
-
-  private
-
-  def msg_valid?(msg)
-    msg =~ /^\d+\ ?\+\ ?\d+$/
-  end
-
-  def add_from(msg)
-    num1, num2 = msg.split('+')
-    result = num1.to_i + num2.to_i
-    
-    "#{msg} = #{result}"
   end
 end
